The file "s3fd.pth" will never load automatically
We check if path_to_detector is None, but `path_to_detector` will never be None because he is hardcoded in method params.
Hence, the file will never be automatically downloaded. I suggest checking if such a file exists and if not, then load it via the "load_url" method.